+++ RELEASE_NOTES.md
@@ -19,3 +19,5 @@
 - Fixed bug with formulas raising exception when stopped.
 
 - Fix a bug that raised `CancelledError` when actor was started with `frequenz.sdk.actor.run` and stopped.
+
+- Stop catching `BaseException` in `frequenz.sdk.actor.run`. Only `CancelledError` and `Exception` are caught now.